Finished Kalman Filter. Model interface changed. Mocks changed.
This commit implements Predict and Update of [Linear] Kalman Filter.

It changed Model interface so that it's underlying matrices are
accessible in order to do some processing on them.

invalidMock in tests now embedds filter.Model instantiate to okModel.
This has a nice effect of removal of unnecessary code from tests.